search bar prototype
